+++ vitest.config.ts
@@ -5,13 +5,15 @@ import react from "@vitejs/plugin-react";
 import tsconfigPaths from "vite-tsconfig-paths";
 
 export default defineConfig({
-  plugins: [
-    react(),
-    tsconfigPaths()
-  ],
+  plugins: [react(), tsconfigPaths()],
   test: {
     environment: "jsdom",
     globals: true,
     setupFiles: "./vitest.setup.ts",
+    coverage: {
+      provider: "istanbul",
+      reporter: ["text", "html"],
+      all: true,
+    },
   },
 });